Started to test the functionality
The system that has been created is being tested. to start this a new Blog template page has been created that  will allow me to view all the data needed for the blog, this also means that in the very near future I can start to properly use it as a Blog and post updates of it's own creation.

Note: All classes and page will need to be reformated in the future. removing useless or redundant code.